+++ Classes/HermesAppDelegate.m
@@ -127,6 +127,7 @@
   [app_defaults setObject:@"1" forKey:PLEASE_GROWL_NEW];
   [app_defaults setObject:@"1" forKey:PLEASE_BIND_MEDIA];
   [app_defaults setObject:@"0" forKey:PLEASE_CLOSE_DRAWER];
+  [app_defaults setObject:QUALITY_MED forKey:DESIRED_QUALITY];
 
   NSUserDefaults *defaults = [NSUserDefaults standardUserDefaults];
   [defaults registerDefaults:app_defaults];