+++ Classes/PreferencesController.h
@@ -12,8 +12,14 @@
 #define PLEASE_GROWL_NEW      @"hermes.please-growl-new"
 #define PLEASE_GROWL_PLAY     @"hermes.please-growl-play"
 #define PLEASE_CLOSE_DRAWER   @"hermes.please-close-drawer"
+#define DESIRED_QUALITY       @"hermes.audio-quality"
+
+#define QUALITY_HIGH @"high"
+#define QUALITY_LOW  @"low"
+#define QUALITY_MED  @"medium"
 
 @interface PreferencesController : NSObject <NSWindowDelegate> {
+  /* General */
   NSButton *scrobble;
   NSButton *scrobbleLikes;
   NSButton *scrobbleOnlyLiked;
@@ -21,6 +27,16 @@
   NSButton *growl;
   NSButton *growlPlayPause;
   NSButton *growlNewSongs;
+
+  /* Playback */
+  NSButtonCell *highQuality;
+  NSButtonCell *mediumQuality;
+  NSButtonCell *lowQuality;
+
+  NSToolbar *toolbar;
+  NSView *general;
+  NSView *playback;
+  NSWindow *window;
 }
 
 @property (nonatomic, retain) IBOutlet NSButton *scrobble;
@@ -30,7 +46,15 @@
 @property (nonatomic, retain) IBOutlet NSButton *growl;
 @property (nonatomic, retain) IBOutlet NSButton *growlPlayPause;
 @property (nonatomic, retain) IBOutlet NSButton *growlNewSongs;
+@property (nonatomic, retain) IBOutlet NSToolbar *toolbar;
+@property (nonatomic, retain) IBOutlet NSView *general;
+@property (nonatomic, retain) IBOutlet NSView *playback;
+@property (nonatomic, retain) IBOutlet NSWindow *window;
+@property (nonatomic, retain) IBOutlet NSButtonCell *highQuality;
+@property (nonatomic, retain) IBOutlet NSButtonCell *mediumQuality;
+@property (nonatomic, retain) IBOutlet NSButtonCell *lowQuality;
 
+/* General */
 - (IBAction) changeScrobbleTo: (id) sender;
 - (IBAction) changeScrobbleLikesTo: (id) sender;
 - (IBAction) changeScrobbleOnlyLikedTo: (id) sender;
@@ -38,5 +62,13 @@
 - (IBAction) changeGrowlTo: (id) sender;
 - (IBAction) changeGrowlPlayPauseTo: (id) sender;
 - (IBAction) changeGrowlNewSongTo: (id) sender;
+
+/* Playback */
+- (IBAction) changeQualityToLow:(id)sender;
+- (IBAction) changeQualityToMedium:(id)sender;
+- (IBAction) changeQualityToHigh:(id)sender;
+
+- (IBAction) showGeneral: (id) sender;
+- (IBAction) showPlayback: (id) sender;
 
 @end